+++ v3/crates/jsonapi/tests/jsonapi_golden_tests.rs
@@ -3,7 +3,7 @@
 use engine_types::HttpContext;
 use hasura_authn_core::{Identity, Role};
 use jsonapi_library::api::{DocumentData, IdentifierData, PrimaryData};
-use metadata_resolve::LifecyclePluginConfigs;
+use metadata_resolve::{LifecyclePluginConfigs, ResolvedLifecyclePreResponsePluginHooks};
 use reqwest::header::HeaderMap;
 use std::collections::{BTreeMap, HashSet};
 use std::path::{Path, PathBuf};
@@ -44,7 +44,7 @@ fn test_get_succeeding_requests() {
 
                 let plugins = LifecyclePluginConfigs {
                     pre_parse_plugins: Vec::new(),
-                    pre_response_plugins: Vec::new(),
+                    pre_response_plugins: ResolvedLifecyclePreResponsePluginHooks::new(),
                     pre_route_plugins: Vec::new(),
                     pre_ndc_request_plugins: BTreeMap::new(),
                     pre_ndc_response_plugins: BTreeMap::new(),
@@ -116,7 +116,7 @@ fn test_get_failing_requests() {
 
                 let plugins = LifecyclePluginConfigs {
                     pre_parse_plugins: Vec::new(),
-                    pre_response_plugins: Vec::new(),
+                    pre_response_plugins: ResolvedLifecyclePreResponsePluginHooks::new(),
                     pre_route_plugins: Vec::new(),
                     pre_ndc_request_plugins: BTreeMap::new(),
                     pre_ndc_response_plugins: BTreeMap::new()