+++ v3/crates/metadata-resolve/src/stages/plugins/types.rs
@@ -2,7 +2,8 @@ use open_dds::{
     data_connector::DataConnectorName,
     plugins::{
         LifecyclePluginName, LifecyclePluginUrl, LifecyclePreNdcRequestPluginHookConfig,
-        LifecyclePreNdcResponsePluginHookConfig,
+        LifecyclePreNdcResponsePluginHookConfig, LifecyclePreResponsePluginHookConfigRequest,
+        OnPluginFailure,
     },
 };
 use serde::{Deserialize, Serialize};
@@ -36,4 +37,63 @@ pub struct ResolvedLifecyclePreNdcResponsePluginHook {
     pub url: LifecyclePluginUrl,
     /// Configuration for the lifecycle plugin hook.
     pub config: LifecyclePreNdcResponsePluginHookConfig,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ResolvedLifecyclePreResponsePluginHooks {
+    /// Synchronous hooks
+    pub sync_hooks: Vec<ResolvedLifecyclePreResponseSyncPluginHook>,
+    /// Asynchronous hooks
+    pub async_hooks: Vec<ResolvedLifecyclePreResponseAsyncPluginHook>,
+}
+
+impl Default for ResolvedLifecyclePreResponsePluginHooks {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ResolvedLifecyclePreResponsePluginHooks {
+    pub fn new() -> Self {
+        Self {
+            sync_hooks: Vec::new(),
+            async_hooks: Vec::new(),
+        }
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.sync_hooks.is_empty() && self.async_hooks.is_empty()
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ResolvedLifecyclePreResponseSyncPluginHook {
+    /// The name of the lifecycle plugin hook.
+    pub name: LifecyclePluginName,
+    /// The URL to access the lifecycle plugin hook.
+    pub url: LifecyclePluginUrl,
+    /// Configuration for the synchronous mode of the plugin hook.
+    pub config: ResolvedLifecyclePreResponseSyncPluginHookConfig,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ResolvedLifecyclePreResponseSyncPluginHookConfig {
+    /// Configuration for the request to the lifecycle plugin hook.
+    pub request: LifecyclePreResponsePluginHookConfigRequest,
+    /// Configuration for the on-plugin-failure behavior of the plugin hook (default: fail).
+    pub on_plugin_failure: OnPluginFailure,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ResolvedLifecyclePreResponseAsyncPluginHook {
+    /// The name of the lifecycle plugin hook.
+    pub name: LifecyclePluginName,
+    /// The URL to access the lifecycle plugin hook.
+    pub url: LifecyclePluginUrl,
+    /// Configuration for the request to the lifecycle plugin hook.
+    pub request: LifecyclePreResponsePluginHookConfigRequest,
 }